# Optimise texture opacity computation algorithm