@peppy 2022-04-04
Did you try checking each corner and the centre pixel as pre-checks? Could potentially speed things up as the majority of transparent images that we use are transparent at once edge (and usually not in the centre).