@smoogipoo 2022-04-04
In the most common of cases (when everything's opaque), you still have to go through every pixel in the image.

Pixels in the top-left and top-right corners are already optimised, and the cases with transparency in corners other than those two are going to be rare so I don't think there's much point to doing that.