@peppy 2022-04-04
Right, that's unfortunate. I guess the only remaining improvement would be to hint at texture load-time? I seem to recall that we discussed and maybe even implemented that in the past...