@smoogipoo 2022-04-04
Yeah, perhaps. I was thinking along the same lines that we could bake some data into the texture cache, but we only use that for fonts right now.