@peppy 2022-04-04
All jpegs could get the opaque flag implicitly, which would cover 99.9% quite well.